refactor(profile/components/common): refactor `StatsOptions` to receive props
